Added unit tests to hit the lines that were modified
